+++ src/config/configure.ts
@@ -11,6 +11,19 @@ export class ConfigHelper {
     this.defaults = defaults;
   }
 
+  /** Verifies the config against some incompatible settings */
+  verifyConfig() {
+    if (this.config.dbServer?.cloudant && this.getItem('session.dbFallback')) {
+      throw 'dbFallback is only implemented for CouchDB.';
+    }
+    if (
+      this.config.dbServer?.iamApiKey &&
+      (this.config.dbServer?.password || process.env.CLOUDANT_PASS)
+    ) {
+      throw 'do not provide a password when using IAM authentication!';
+    }
+  }
+
   getItem(key: string) {
     let result = getObjectRef(this.config, key);
     if (typeof result === 'undefined' || result === null) {